+++ serverless/CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.1.0] - 2023-09-06
+
+### Added
+
+- encoding template 에서 hls 인코딩 작업이 추가됩니다.
+
+### Remove
+
+- 더이상 encoding lambda 함수에서 Settings으로 인코딩 옵션을 조정하지 않습니다.
+
 ## [1.0.0] - 2023-08-30
 
 ### Added